Automatic winsorization (#722)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/722

This adds support for automatically inferring the winsorization settings based on the observed data and the optimization config, which is useful for outcome constraints and MOO. The main idea is to winsorize in a similar way to how outliers are defined in a Tukey box plot, but while also taking the constraint boundaries into account. For users fitting models in a notebook where there is no optimization config it is also possible to use automatic winsorization by specifying, e.g., `WinsorizationConfig(0.0, AUTO_WINS_QUANTILE)`, which will automatically winsorize from above and apply no winsorization from below.

When an optimization config is supplied, the following logic is used:
- **Single objective:** We automatically winsorize from above for minimization and from below for maximization. For minimization the cutoff is `q3 + 1.5 (q3 - q1)`  and for maximization it is `q1 + 1.5 (q3 - q1)`  where `q1` is the first quartile and `q3` is the third quartile.
- **Outcome constraint:** For a leq constraint, we will winsorize from above but need to be careful so we don't make infeasible points feasible. Thus, instead of using the cutoff `q3 + 1.5 (q3 - q1)` we instead use `max(q3, bnd) + 1.5 (max(q3, bnd) - q1)`. Using the max of the 3rd quartile and the bound guarantees that infeasible points remain infeasible.
- **Multi objective:** If an objective threshold is specified, we deal with this in the same way as an outcome constraint. If no objective threshold is specified we don't winsorize (since outliers may be on the Pareto frontier).

Automatically infering winsorization from the optimization config doesn't support:
- Relative outcome constraints (including relative objective thresholds)
- Scalarized objectives
- Scalarized outcome constraints.

Reviewed By: bernardbeckerman

Differential Revision: D32306496

fbshipit-source-id: 5bccc376c8d6eae2cf3c1b65a0253083e243ea96