Change existing MapMetrics to MilesMapMetrics (and fix all associated tests)
Summary:
Apologies for the mega-diff.

This diff does a number of things: Create the new MilesNoisyFunctionMapMetric, change existing noisy map functions to be children of this new class (and fix associated tests), and change early stopping to accept MilesMapMetrics (and fix associated tests). Experiment is also significantly changed to pass around MilesMapData appropriately.

Reviewed By: bernardbeckerman

Differential Revision: D32186705

fbshipit-source-id: 8799cf9e4104e9d9fec55bb402626d8e0e77bc83