Added new .true_df field
Summary:
Added a new field to return the df acting as the source of truth for all Data types. This is not the cleanest thing (as we have no guarantees on the shape of this dataframe) and on principle I think we should avoid using it when possible, but it becomes necessary to avoid any implicit casting in our caching layer.

With this, our cache will allows capture maximum availible information.

Reviewed By: bernardbeckerman

Differential Revision: D32187153

fbshipit-source-id: dd49dbf7df10e16938ed73fb1fd39b192aa04ecd