Delete MapData
Summary:
Its finally happening! This diff deletes map_data.py and chases down all the errors. Most are small and fairly straightforward (a big one was making the default map_key_info value 0.0 (float) instead of 0 (int) in a number of places). As of this diff only MilesMapData is used in Ax.

Next step: Rename MilesMapData to MapData.

itshappening

Reviewed By: bernardbeckerman

Differential Revision: D32504462

fbshipit-source-id: 7f1c9095125c74375a5e01d17842aa67e167a9a4