Update docstring for AxClient.update_running_trial_with_intermediate_data.
Summary:
Adds an explanation for what happens if you pass the same `raw_data`
multiple times and how to get around it.

I ran into this issue when I was putting together N1403208 and it was unclear
what the issue was or how to fix it (without looking at e.g. tests).

Reviewed By: lena-kashtelyan

Differential Revision: D32761138

fbshipit-source-id: e7392b98bceff9227f6da2958de1e1bcc09c3e69