attach_data bug fix
Summary:
We've been running into an issue where under a very specific set of circumstances an experiment's data cache will become incorrectly overriden and we will lose a metric.
We noticed this when looking up data on an AxSweep run (where the objective was available while running but the tracking metric was not) came back without its tracking metric, but fetch data got all metrics perfectly fine.
The issue was that a bit lof logic in attach data was incorrectly overriding the cache: There are some situations in which we are confident that the newly fetched data is entirely overlapping with the cached data and in this case we deem it safe to override the cached data. The logic for determining this was broken (...and writen by me on my first week with the team <3) but now fixed!

Reviewed By: ldworkin

Differential Revision: D33168874

fbshipit-source-id: 42835a66cb54e82f5190861c0a66e2a836366d67