Declaratively select model specs in analyzers
Summary:
This diff breaks out a bunch of abstractions that were part of analyzer.

There are `ModelSpecCaseEvaluator` and `ModelSelectorCaseEvaluator` which based on a  `GenerationNodeCase` know whether they should be used.  `GenerationNodeCase.build` can create a case from the analyzer's child objects.

`GenerationNodeBuilder` has a list of both kinds of evaluators, and can produce generation steps from ax objects.

`ModelSpecBuilder` is responsible for returning a copy of a given modelspec with the PTS defaults set.

Rename to follow, because we refer to the `PTSGenerationStep` as a node a lot in this diff.

Reviewed By: ldworkin

Differential Revision: D33197938

fbshipit-source-id: 713ae56709d360fc1d1b2e5b3f26303a295eb2a2