Plot objective vs true_objective in get_standard_plots
Summary:
Plot the true_objective_metric against the experiment's objective when an early stopping with true objective is supplied.

For context, some teams may use a surrogate metric that is available while running for early stopping that they believe correlates with their true objective, and at the end of their experiment would like to visually confirm the correlation.

Reviewed By: bernardbeckerman

Differential Revision: D33026053

fbshipit-source-id: 40dde01e17e5453658ef2b892acd6356a9fec103