Make models enum non-optional for `get_model_from_generator_run`
Summary: This being optional was creating a bug-prone codepath due to our use of the method on generator runs using internal only models. This change to the models enum being required cascaded through a fairly large portion of our code, stemming from our handling of best_point functions, but I think it will be worth it and not as disruptive as one may think.

Reviewed By: lena-kashtelyan

Differential Revision: D33238024

fbshipit-source-id: ce3bd7089ad208f19c3b88d7b5d22483e2b59986