relax Winsorization warnings tests
Summary: Previous tests required specific warnings in a specific order, whereas all we really want to be sure of is that all specified warnings are present. This relaxes that requirement so that the tests can pass and the oss package can release. This could also be fixed by pinning the numpy version in oss to <1.22.

Reviewed By: lena-kashtelyan

Differential Revision: D33513162

fbshipit-source-id: 3a77183ef76249d576a5369b6e49568981341f30