Return a useful error when cross validate gets called on SOBOL
Summary:
I stumbled upon this through a bug when developing: CV got called on SOBOL, though it shouldn't have, which lead to a `ZeroDivisionError`.  This creates more of a challenge to debug than it should because the user doesn't necessarily know it's a SOBOL model in play where the error is thrown.
It turns out the bridge's CV `NotImplementedError` never gets thrown because CV is only called on the training data.  Rather than force it to, I thought it was enough to just name the model.  Plus, no training data would still be a problem for GPEI

Reviewed By: EugenHotaj

Differential Revision: D33351632

fbshipit-source-id: 42df3cae58df8bf2e8d29822dad3d7d5ae8269d7