Turn off relativization in objective plots
Summary:
The true objective metric may not have associated data in the status_quo are so trying to create a relativized plot will always fail, resulting in the plot being missing from the report in FBLearner.

We turn off relativization to avoid this.

Also changed the way the objective graph sources its Y-axis label and whether or not to minimize/maximize to address visual mismatch bugs that were noticed while fixing the primary issue.

Reviewed By: bernardbeckerman

Differential Revision: D33482616

fbshipit-source-id: b48c5070efc889151819419f703294b13dddc1bc