Add decorator to inject `maxiter=1` into scipy.optimize.minimize
Summary:
HUGE TESTING WIN!!

This is an evolution of the work started in D33664258, with a big idea from EugenHotaj to mock out one layer lower in botorch. Instead, we now mock out two layers lower, changing calls to scipy's `minimize` in botorch's fitting and candidate generation.

This diff adds a decorator that when active, will force scipy the spend less time in the minimize function. When applied to a test (as I have in the infamously slow test_factory.py) the models come out with poor performance (which we should not care about in unit tests) but otherwise get fully tested! This should not affect coverage at all.

Reviewed By: EugenHotaj

Differential Revision: D33714519

fbshipit-source-id: 00b554905a2d4bfa60bad801854774ccf1ccfee9