Add new fast_botorch decorator to everywhere our test suite has a non-SOBOL model
Summary:
Add the new decorator to a ton of tests! In all likelihood this is not exhaustive, but its anywhere BigGrep found we used a non-Sobol model in a file name that matched `ax/ax.*test`.

I haven't tested the speedup of each test individually but we're seeing ~60s tests go town to ~20s. Once this result lands Im certain we'll see a huge anti-spike in the dashboard here https://www.internalfb.com/intern/unidash/dashboard/ae_eng_devx/home/

Reviewed By: lena-kashtelyan

Differential Revision: D33719604

fbshipit-source-id: 435c356c40d2df5b02ae39ae265cc4eb31183074