Pass candidate metadata when extracting pending features
Summary:
When we create pending features, we should also add candidate metadata to them, since that metadata can impact the form, which pending features should ultimately take when passed to the model.

For example, for HSS we inject parameters we removed from the arm before returning it to the user (because those parameters came from flattened version of the search space and didn't make sense given the hierarchical structure of the search space), into candidate metadata, before passing that arm back to the model (we do this so the model gets back as pending the exact arm that it generated).

Reviewed By: Balandat

Differential Revision: D33520367

fbshipit-source-id: bd98b2a0c75c735637884762ac65bb11e4c56bf7