Add .gen() call to all modelbridge tests
Summary:
Improving test coverage. In doing this I've found MTGP may be broken! I've taskified fixing this and have a repro notebook N1548415 but I may not be the best person to look into this (hit a bit of a wall).

If we are satisfied with this we may be able to turn off cogwheel tests at diff time?

Reviewed By: EugenHotaj

Differential Revision: D33853623

fbshipit-source-id: 33308f2861d13324343fc6bdb8483071392b5137