Getting rid of metric data fetching properties
Summary:
It doesn't make sense to specify these properties on metrics, because we attach data for *multiple* metrics at once. So if one metric says to overwrite and the other doesn't, what do we do? To handle this properly we'd need some really complicated way of updating old dataframes in piecemeal, which we don't want to be in the business of right now.

We're not even using combine_with_last_data, and the only reason we were using  overwrite_existing_data was to save space in the database. Let's just kill these for now -- it won't break anything, and the database issue will be addressed in the next diff.

Reviewed By: lena-kashtelyan

Differential Revision: D33277898

fbshipit-source-id: e5640dd9267f6be5a0775b9df6708ab76ad5cbba