Moving attach_data to from metric.fetch_data to experiment.fetch_data
Summary:
A lot of the confusion around our lookup/fetch/attach data flow is due to the following behavior.

A side effect of calling `fetch_data` is that we also call `attach_data` from within this function. This is desirable for a number of reasons and we don't want to get rid of this behavior. However, the time and place we were calling `attach_data` was very nonintuitive.

The pseudocode of experiment.fetch_data is something like this:
```
def fetch_data():
   for metric in exp.metrics:
      metric_data = metric.fetch_data()
      data_list.append(metric_data)
   return pd.concat(data_list())
```
The pseudocode of metric.fetch_data is something like this:
```
def fetch_data():
   new_data_for_this_metric = talk_to_external_service()

   data_for_all_metrics = exp.lookup_data()
   updated_data_for_all_metrics = pd.concat(
      new_data_for_this_metric,
      data_for_all_metrics,
   )
   exp.attach_data(updated_data_for_all_metrics)

   return new_data_for_this_metric
```

This is so confusing! We're calling exp.attach_data from within the *metric specifying* data fetching code, which requires us to have a lot of complicated logic, and results in duplicates and/or things getting overridden, etc etc.

A better system would be to move all the data_attaching to experiment.fetch_data, i.e.:

```
def experiment.fetch_data():
   for metric in exp.metrics:
      metric_data = metric.fetch_data()
      data_list.append(metric_data)
   data = d.concat(data_list()
   exp.attach_data(data)
   return data

def metric.fetch_data():
   new_data_for_this_metric = talk_to_external_service()
   return new_data_for_this_metric
```

The reason we're not doing this is due to the combine_with_last_data and overwrite_existing_data flags, but I think we can figure out a better way to handle that.

Reviewed By: lena-kashtelyan

Differential Revision: D33267841

fbshipit-source-id: 60bf2b73944b91bd0b890c561900329bcbad81aa