- Add test for typechecking OptimizationConfig
Summary:
Add a test that catches when a warning is thrown after a bad parameter is passed to a Model. This test will be removed after TConfig is updated, but adding here for now.

V2 - contains test that catches warning on OptimizationConfig
V3 - Removes warning catch in unit test, migrates TConfig, and adds OptimizationConfig as a valid type.

Reviewed By: lena-kashtelyan

Differential Revision: D34086883

fbshipit-source-id: 10c07791f696384f29d307c1133effdb348204e0