Add optional optimization config to best point methods
Summary: Add the ability to override the optimization config present on the experiment in best point extraction. Immediate use case relates to fetching the best point based on a tracking metric (the true objective metric from IGML's use case) but this will likely be relevent to other users as well.

Differential Revision: D34400863

fbshipit-source-id: d3e7fc1641f0d8c7ab9aa27d55eea0a86b0b0799