Improve `@fast_botorch_optimize` by reducing IC generation cost
Summary: Using `fast_botorch_optimize` implies that we do not care about the optimization quality, in which case we do not care about the quality of the initial conditions used to start the optimization either. This appends the `fast_botorch_optimize` context manager to overwrite the `num_restarts` and `raw_samples` with `2` and `4` respectively, significantly reducing the computational effort that goes into generating the initial conditions.

Differential Revision: D34432833

fbshipit-source-id: 590dbbddc8d15ace1ee63e5f6c4332f80656dd0c