Improve `@fast_botorch_optimize` by reducing IC generation cost (#820)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/820

Using `fast_botorch_optimize` implies that we do not care about the optimization quality, in which case we do not care about the quality of the initial conditions used to start the optimization either. This appends the `fast_botorch_optimize` context manager to overwrite the `num_restarts` and `raw_samples` with `2` and `4` respectively, significantly reducing the computational effort that goes into generating the initial conditions.

Reviewed By: mpolson64

Differential Revision: D34432833

fbshipit-source-id: 0dd1b5022b351c1be4f0d4e75ad04dbb7ec187c9