Add optional optimization config to best point methods
Summary: Add the ability to override the optimization config present on the experiment in best point extraction. Immediate use case relates to fetching the best point based on a tracking metric (the true objective metric from IGML's use case) but this will likely be relevent to other users as well.

Reviewed By: lena-kashtelyan

Differential Revision: D34400863

fbshipit-source-id: c112af0185c885c8d230237c806d30f9df23f0e3