Use Scheduler.best_point in param_sweep_scheduler (#818)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/818

This diff is a step towards deprecating the best point extraction from report utils.

In Ax we have two ways of extracting the best point from an experiment, one in the report utils and another in the BestPointMixin. The latter is much better supported, notably having the ability to find the best point based on model predictions and respect outcome constraints, and ideally we should stop using the one in report utils.

This diff removes the call to ax.service.utils.report_utils.get_best_trial and replaces it with Scheudler.get_best trial and some glue code. AFAIK ax.service.utils.report_utils.get_best_trial is only called in some tests now so it should be safe to delete, but just in case I have only marked it as deprecated.

Reviewed By: lena-kashtelyan

Differential Revision: D34420658

fbshipit-source-id: 472c9a2f74293d4a03a50438e64b42c3192e22bc