Fix metric encoding flow to allow for more descriptive error to surface
Summary:
A user brought this to our attention when they tried to save their experiment but forgot to use the PTS registry bundle and received an undescriptive KeyError instead of the message pointing them to the post on registry changes.

This fix (using .get instead of indexing) is already in place on Runner encoding and must have just slipped through the cracks here.

Reviewed By: lena-kashtelyan

Differential Revision: D34642585

fbshipit-source-id: 5308d058c09ffec87df12824e8c7658e0db27fbf