docs: add social banner in support of Ukraine (#832)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

<img width="1080" alt="image" src="https://user-images.githubusercontent.com/12485205/156690612-b9404b32-13ee-4f06-b5c6-b4716b5de46c.png">

Pull Request resolved: https://github.com/facebook/Ax/pull/832

Reviewed By: mpolson64

Differential Revision: D34633533

Pulled By: dmitryvinn-fb

fbshipit-source-id: e7f66bab2a1f9907a79da3d1ae12d68cdfe63846