Raise an informative warning on call to register_{metric, runner}
Summary: As we all know these methods have undergone a bug change that may not be intuitive to everybody. Because the methods retained the same name through the refactor but do relatively different things, we should log a warning whenever this method is used (at least for now) to help our OSS users. This message links to the appropriate section of a tutorial that they should find helpful (the same link is provided on metric/runner encoding/decoding failure.

Differential Revision: D34732470

fbshipit-source-id: 9e42017851143948beeaa4ddaead9c8f1fd0e927