Raise an informative warning on call to register_{metric, runner} (#840)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/840

As we all know these methods have undergone a bug change that may not be intuitive to everybody. Because the methods retained the same name through the refactor but do relatively different things, we should log a warning whenever this method is used (at least for now) to help our OSS users. This message links to the appropriate section of a tutorial that they should find helpful (the same link is provided on metric/runner encoding/decoding failure.

Reviewed By: lena-kashtelyan

Differential Revision: D34732470

fbshipit-source-id: 230b0a655dfda2fb2ef80a2e4c6c69381981e6e2