BenchmarkProblem and from_botorch
Summary:
Create a new BenchmarkProblem class containing a SearchSpace, OptimizationConfig, and Runner.

Also create two subclasses for single and multi objective problems (the single objective problem has a known optimum, the multi has a known max hypervolume).

Included are factory methods to turn botorch test functions into BenchmarkProblems. In these functions the search space is sized to match the bounds of the botorch function and the optimization config is made to have metrics that can extract the value of the .forward method from the runner, which calculates .forward. I chose to do this (as opposed to evaluating the synthetic function on metric) because for the multi objective case the .forward method computes all metrics in one tensor; it didnt make sense to calculate this same tensor over and over again.

Differential Revision: D34827595

fbshipit-source-id: c448c11055ffa295358a90e52dd351d6cafb9869