BenchmarkProblem and from_botorch
Summary:
Create a new BenchmarkProblem class containing a SearchSpace, OptimizationConfig, and Runner.

Also create two subclasses for single and multi objective problems (the single objective problem has a known optimum, the multi has a known max hypervolume).

Edit for clarity: users can instantiate BenchmarkProblems and any of their subclasses without the from_botorch method. This will be useful for real world problems where we will construct the search space, optimization config, and runner directly. The from_botorch factory methods are just a convenience method on top of the regular dataclass initializer for a use case I forsee being relatively common.

Included are factory methods to turn botorch test functions into BenchmarkProblems. In these functions the search space is sized to match the bounds of the botorch function and the optimization config is made to have metrics that can extract the value of the .forward method from the runner, which calculates .forward. I chose to do this (as opposed to evaluating the synthetic function on metric) because for the multi objective case the .forward method computes all metrics in one tensor; it didnt make sense to calculate this same tensor over and over again.

Reviewed By: lena-kashtelyan

Differential Revision: D34827595

fbshipit-source-id: feb8672773813d2845650ed6883209ce72776144