Introduce a class for specifying parameter distributions (#867)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/867

Introduces a `ParameterDistribution` class for specifying a parameter distribution. This takes in a `scipy` distribution name and the arguments needed to construct the distribution. The distribution is imported and constructed dynamically.

The overall idea is to use this to specify the parameter distributions, which we would sample from and use in constructing the `InputPerturbation` or `AppendFeatures` input transforms, which would be passed onto the BoTorch model.

Reviewed By: Balandat

Differential Revision: D34943939

fbshipit-source-id: a3072c2bcb88d7dcdd3819a923a3bc4fc99c76e9