Move SchedulerOptions so they can be stored
Summary:
Unfortunately, adding SchedulerOptions to the json registry created a circular dependency because of things imported by the Scheduler. Since it is a pretty lightweight dataclass there is no need for it to necessarily have to live in the same module as the Scheduler so I have popped it out into its own module in util. I have done some aliasing in Scheduler so users can still import it from there.

This doesn't feel like a great solution but I'm not sure if there are any better options. Open to all ideas

Differential Revision: D35147597

fbshipit-source-id: 98294bb084a1303fd0251c25cf8062377cf1781d