Make Benchmark objects json storable (#887)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/887

Add json storage support for the BenchmarkProblem, BenchmarkMethod, BenchmarkResult, and AggregatedBenchmarkResult so they can be passed around as parameters and results on fblearner.

This wound up requiring a little bit of TARGETS rewriting in the benchmarking suite to keep Scheduler dependencies out of the registry, since this would create a dependency loop.

Reviewed By: lena-kashtelyan

Differential Revision: D35146551

fbshipit-source-id: f0b189474eda7924297e89ecc88d57514426e013