Move SchedulerOptions so they can be stored (#886)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/886

Unfortunately, adding SchedulerOptions to the json registry created a circular dependency because of things imported by the Scheduler. Since it is a pretty lightweight dataclass there is no need for it to necessarily have to live in the same module as the Scheduler so I have popped it out into its own module in util. I have done some aliasing in Scheduler so users can still import it from there.

Also, changed trial_type to use an enum instead of the raw class. This makes serialization much easier and has the added benefit of not forcing us to carry around the Trial import wherever the SchedulerOptions goes.

Reviewed By: lena-kashtelyan

Differential Revision: D35147597

fbshipit-source-id: 4f03d67a801d0018dc77747ccb8c542701424e0a