Implement QuickBO
Summary:
This is just a bare bones, no change in interface, just if the user adds a long run trial, the client treats it as the experiement as quickbo as long as the long run trial is running.
We've upstreamed `instagram.igml.assistant.qe_lib.utils.analyzer.QELibQuickBOAnalyzer` to `ax.fb.utils.pts.QuickBOAnalyzer`, minus the `QELibTrialBuilderMixIn`.  I'm going to leave the tests on `QELibQuickBOAnalyzer` to make sure the contract is preserved as there are users using that one, and for PTS I eventually want to de-differentiate the sequential and quickbo analyzers.

Reviewed By: lena-kashtelyan

Differential Revision: D35293585

fbshipit-source-id: 2fb436f5897630b6579a034e464bd08c2f359138