ScoredBenchmarkResult dataclass (#902)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/902

Create a dataclass for judging performance of a benchmark method vs some baseline. The result is calculated as described in appendix B https://arxiv.org/pdf/2104.10201.pdf where 0 means as good as the baseline and 100 means true optimum found.

Also included is a function for generating these baseline results, using 100 replications of a 100-trial sobol method (again, open for input on these parameters). We include so many trials as a method of future proofing (ie we can use this baseline to compare against and method using up to 100-trials). Long term, the plan will be to run this baseline method for all our built-in problems and dump the output to JSON so we can produce normalized results whenever we need without re running sobol trials.

Reviewed By: dme65

Differential Revision: D35372570

fbshipit-source-id: 0e9d0f86f6400e4430956671b4a32dd81d6635a9