Update transforms to support RobustSearchSpace / ParameterDistribution
Summary:
This adds support for transforming `ParameterDistribution`s to `UnitX` and `CenteredUnitX`, and raises errors in a number of other transforms if they attempt to transform a parameter with an attached distribution.

This makes `CenteredUnitX` a subclass of `UnitX` with added support for constraints.

Also addresses a couple minor comments on `RobustSearchSpace`.

Reviewed By: Balandat

Differential Revision: D35439048

fbshipit-source-id: 484bee6a1fbd1317b6ace2306d092cbeb0649b0f