Update transforms to support RobustSearchSpace / ParameterDistribution (#904)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/904

This adds support for transforming `ParameterDistribution`s to `UnitX` and `CenteredUnitX`, and raises errors in a number of other transforms if they attempt to transform a parameter with an attached distribution.

This makes `CenteredUnitX` a subclass of `UnitX` with added support for constraints.

Also addresses a couple minor comments on `RobustSearchSpace`.

Reviewed By: Balandat

Differential Revision: D35439048

fbshipit-source-id: 2661cde261106c779b5a044a32a08cd85002c322