Create nightly fblearner workflow
Summary: Create a workflow which will run all our most used methods on all the relevant test problems and in the future log their results to a scuba table. For now it just runs MBO GPEI on Branin and Ackley functions but we can add more methods and problems in the near future.

Reviewed By: lena-kashtelyan

Differential Revision: D35400834

fbshipit-source-id: 017dc91ecb21feb2fa0a24be76a36f6ebb548d66