Support early stopping on non-objective metrics (#911)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/911

It is not hard to imagine wanting to early stop on some non-objective metric, say a guardrail metric. This, along with the new LogicalEarlyStoppingStrategy, will allow users to create extremely expressive early stopping strategies that capture all their preferences for a worthwhile solution.

Metric names are passed in as an iterable for compatibility with some ideas Daniel had for future strategies that may early stop on multiple metrics in more complex ways than LogicalEarlyStoppingStrategies provide, talk to him for more details

Reviewed By: lena-kashtelyan

Differential Revision: D35549557

fbshipit-source-id: a4677d18fe127e2b85c2da9b35e20aefb8f92c7b