Update SearchSpaceDigest to support RobustSearchSpace
Summary:
Adds `environmental_variables` and `distribution_sampler` attributes to `SearchSpaceDigest`. The `distribution_sampler` is a callable that generates a `num_samples x d`-dim tensor of samples, where `d` is the number of environmental variables in the case of environmental variables and the number of parameters in the case of input perturbations.

We can use the `distribution_sampler` to generate either the `perturbation_set` or `feature_set`, which would be used to construct the `InputPerturbation` or `AppendFeatures` transform.

Reviewed By: Balandat

Differential Revision: D35516875

fbshipit-source-id: b794de58a8322250bcfa24bf5d015958e3ccc3f3