Add method to dynamically created BenchmarkMethods via choose_generation_strategy (#917)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/917

Of course it is very important that we are able to observe the performance of choose_generation_strategy over time, but it is a bit of a special case: It is dynamic on aspects of the BenchmarkProblem. This is not a prohibative problem, we can pass in the problem and generate the method using its search space and optimization config, but some special casing needs to be done for full_run. This diff should cover all that special casing.

Reviewed By: lena-kashtelyan

Differential Revision: D35686636

fbshipit-source-id: ea845df07435d94e33480d056274afc7e3125ab7