fix TorchModelbridge._untransform_objective_thresholds for MOO
Summary:
# Context
TorchModelbridge requires MOO model to include all the metrics in `experiment.optimization_config` before model fitting. Otherwise, if users pass a optimization_config with a metric not in  `experiment.optimization_config`  to `model.gen`,  it raises a key error even though the metrics are included in the data passed to the modelbridge. Check N1802207 for repro. It is a pain for users to refit the `modelbridge` in particular when the gp model fitting is expensive (e.g. fully bayesian).

{F723595510}

 This is uses modelbrige._optimization_config in the untransform. Therefore, if users update optimization_config input to `modelbrige.gen`, it will raise keyErrors.

# Mitigation
Pass metrics list of the optimization config in `model.gen` to `_untransform_objective_thresholds` .

Reviewed By: Balandat

Differential Revision: D35711527

fbshipit-source-id: aa0c40dc8a52c16b4a3ff80b5c509ac511acad51