Update outdated Sobol dimension error message (#924)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/924

The current behavior looks for hard-coded `1111` in the error message. This is outdated as the current `MAXDIM` is `21201`.

I simply removed the hard coded dimension from the error message. An alternative would be to get rid of the `make_and_optimize_acqf` helper and the error handling around it. I'm happy to do that as well. Left it in here, in case someone tries to use SAASBO in 25000 dimensions (I'm not even sure if SAASBO uses this part of the code...)

Reviewed By: Balandat

Differential Revision: D35763023

fbshipit-source-id: 0f0b4b78f2b66e5bf1bc4dab1dd873d06eb71167