Modify transforms to support `RobustSearchSpace`, part 2 (#927)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/927

This fixes the ones I originally marked as ignore / no-op and left untouched. Seeing my `RobustSearchSpace` mysteriously being converted to `SearchSpace` made me realize that these needed modifications to respect the original search space class.

Also removes a few pyre-ignore comments.

I also noticed that `RobustSearchSpace` did not complain if we included the environmental variables in both the `environmental_variables` and the `parameters`, which we were doing with some of the modified transforms. This did not seem like a great idea, so it also got fixed.

Reviewed By: Balandat

Differential Revision: D35797496

fbshipit-source-id: a1dce72d87e4db16961b9e01738dcfbf2629d548