Remove hacky fix for pytorch tolist issues
Summary: Platform010 strikes back! The lines removed were added as a quick fix to avoid a sketchy codepath in pytorch where tensor.tolist could throw an UndefinedBehaviorSanitizer error. The error was blocking push so I added this little check to dodge the problematic method, but now that pytorch issue is fixed  D35971229 and we don't need the hack anymore

Reviewed By: lena-kashtelyan

Differential Revision: D36007585

fbshipit-source-id: 49f9cbc5930dc28b73ea312e9ff88be9f1e1e056