Fix bug in observed Pareto plot with relative objective thresholds
Summary:
The observed Pareto plot utility allows you to specify a list of arm names to be included in the plot, in case you want to plot only a subset of the Pareto front. This is accomplished by filtering the data to just the armes included in that list. This causes an issue if the experiment has relative objective thresholds, and status quo wasn't included in that list, because now it has been removed from the data and the utility isn't able to derelativize the objective thresholds.

This diff fixes the bug by ensuring that the status quo is always included in the plot, if there is one specified on the experiment.

Reviewed By: cpueschel

Differential Revision: D36075043

fbshipit-source-id: 9817680900c7fb772805280b9c869e91235b7c6b