Filter data to allow conversion to block design in MBM (#970)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/970

In some situations it may happen that we try to convert data that does not adhere to a block design to a block design representation. For instance, this can happen when manually specifying a `Surrogate` (not `ListSurrogate`) in MBM in a situation where different metrics have a different number of observations (e.g. due to different fetching logic).

Reviewed By: dme65

Differential Revision: D36383355

fbshipit-source-id: 324f43f3c83326fbbf6241df69ca0ad35829febb