fix dummy example that produces identical job_ids for different trials (#952)
Summary: OSS issue #952 reported seeing multiple separate trials with identical paramters when using a Sobol generation strategy. This was due to the scheduler example using timestamp in seconds as the job_id, so that if multiple jobs were scheduled in the same second, they would all have the same ids and therefore there would not be internally distinguished. This fixes by using timestamp microseconds instead, which will be much more reliably different from one scheduled trial to the next.

Reviewed By: lena-kashtelyan

Differential Revision: D36547075

fbshipit-source-id: c0db14b80fb8afd8f964e0fe79ee07f2eee4725f