New scoring algorithm
Summary:
After speaking with Said and David we've decided to change the scoring system one more time.

Now, the best point from the first generation step (in the nightly run always 5 sobol steps) is used as the baseline for the replication. This prevents the issue we had with the old method where a single lucky sobol run in the baseline would cause methods to have wildly negative scores on the aggregated result and SEM so high the results were effectively unusable. This also prevents us from having to save baseline results.

Differential Revision: D36605946

fbshipit-source-id: 7c7435b78a16027d81546dc1151c9338bf61f41d