Add optional tracking metrics to BenchmarkProblem
Summary:
This allows users to specify non-objective metrics to be attached the the experiment and calculated during a benchmark run.

Some changes were made to the init methods on BenchmarkProblem and its subclasses to allow for default arguments in inherited dataclasses, a known shortcoming of the dataclass concept that has been remedied in 3.10 via kw_only.

Reviewed By: danielrjiang

Differential Revision: D36673344

fbshipit-source-id: ad770bb3f7f58720b9c2e8a946815df9ac8be7f6