fix ListSurrogate.construct for constructing MTGP (#1246)
Summary:
X-link: https://github.com/pytorch/botorch/pull/1246

Pull Request resolved: https://github.com/facebook/Ax/pull/994

We see an error in CV when using ListSurrogate (a list of MTGP) in Models.BOTORCH_MODULAR. See N2054031 for the full error trace.

The error happens when loading state_dict. The reason is that a wrong task_feature is passed to MTGP (use default value = 0) and thus, the task_covar_module has a wrong shape.

The proposed the fix is
1.  ListSurrogate.construct. Pass task_feature to contruct MTGP. Also add checks to make sure that len(task_features) = 1
2. Make task_feature as required input in the MultiTaskGP.construct_inputs

Reviewed By: Balandat

Differential Revision: D36925653

fbshipit-source-id: 4ce99abdaba28a5404ab540568b57ec43af3f188