fix ListSurrogate.construct for constructing MTGP
Summary:
Motivation:
We see an error in CV when using ListSurrogate (a list of MTGP) in Models.BOTORCH_MODULAR. See N2054031 for the full error trace.

The error happens when loading state_dict. The reason is that a wrong task_feature is passed to MTGP (use default value = 0) and thus, the task_covar_module has a wrong shape.

The proposed the fix here is
1.  ListSurrogate.construct. Pass task_feature to contruct MTGP. Also add checks to make sure that len(task_features) = 1

Reviewed By: Balandat

Differential Revision: D36946435

fbshipit-source-id: 5c96aafd0849a57c4ceb4bfd740048c4d049da5c