Shuffle around DAG in nightly run
Summary:
This change in DAG structure should address many if not all of the OOM issues we've been seeing in FBLearner. Runs like f352007573 were failing due to ComputeLeafFutures running out of memory, which seems to happen only when outputs of the full workflow are too large for FBLearner to handle. Our output was the aggregated result of all benchmark tests (experiment count in the low 5 figures)!

This diff doesnt change core functionality, but swaps some things around: First, we log results to scuba as we go. This prevents us from needing to grab all the results at once to log. Second, instead of outputing AggregatedResults, we output a dict of `(problem_name, method_name) -> operator_id`. This will still allow for manual inspection of results in bento, but shrinks the workflow output massively.

Reviewed By: bernardbeckerman

Differential Revision: D37314435

fbshipit-source-id: 77e3ed553b932c12d22223fc86dfdeeadcb64c8d