disable profiler for test_runs_detail
Summary:
```
  1) automl.internal.searcher.base.tests.test_base.SearcherTest: test_runs_detail
    1) AssertionError: 1 is not None
      File "automl/internal/searcher/base/tests/test_base.py", line 89, in test_runs_detail
        self.assertIsNone(loop.searcher.average_run_cost)
      File "/usr/local/fbcode/platform010/lib/python3.8/unittest/case.py", line 1322, in assertIsNone
        self.fail(self._formatMessage(msg, standardMsg))
      File "/usr/local/fbcode/platform010/lib/python3.8/unittest/case.py", line 753, in fail
        raise self.failureException(msg)
```
failures started around when D37323868 landed.  My theory is that code runs about 1% slower with the profiler, which creates a tangible "run_cost" here.  It only fails most of the time with profiling turned on.

This also introduces `disable_profiler` which can disable profiling for a specific test, instead of the test case.

Differential Revision: D37417125

fbshipit-source-id: da18e26aca6df78cf2a0b0c878071beba6f45585