Do not allow serialization on manually constructed Surrogates
Summary: This solves a bug we were seeing in the Surrogate benchmark problem I constructed for David. Before this when we tried to serialize a manually constructed Surrogate we would not have all the data needed for serialization, but instead of failing we would simply silently pass in default arguments. Then at decode time we would successfully rebuild the Surrogate from these default arguments, resulting in a totally different model! This will prevent silent errors like this from happening again.

Reviewed By: lena-kashtelyan

Differential Revision: D37323065

fbshipit-source-id: 31d74faf42ddef229594d68cb8b94d5c0cd040bb